UI Stats finalmente responde

+++ Assets/Scripts/UI/UIStatsHandler.cs
@@ -11,13 +11,14 @@ public class UIStatsHandler : MonoBehaviour
 
     public void UpdateHealth(int currentHealth, int maxHealth)
     {
-        healthSlider.value = currentHealth / maxHealth;
+        healthSlider.value = (float)currentHealth / (float)maxHealth;
         healthText.text = $"HP\n{currentHealth} / {maxHealth}";
     }
 
     public void UpdateMana(int currentMana, int maxMana)
     {
-        manaSlider.value = currentMana / maxMana;
+
+        manaSlider.value = (float)currentMana / (float)maxMana;
         manaText.text = $"MANA\n{currentMana} / {maxMana}";
     }
 }
